requirement.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
faiss-cpu==1.9.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
img2pdf==0.5.1
Jinja2==3.1.4
jiter==0.6.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ocrmypdf==16.4.1
openai==1.52.2
orjson==3.10.5
packaging==24.1
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==0.17.0
pikepdf==9.0.0
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.6.0
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.16.0
